Move register struct to its own file
